[Security Solution] render threat intelligence overview new document flyout in Security Solution and Discover
