+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/threat_details_view_enrichment_accordion.tsx
@@ -21,7 +21,7 @@ import { css } from '@emotion/react';
 import { i18n } from '@kbn/i18n';
 import { THREAT_INTELLIGENCE_ENRICHMENTS_ACCORDION_TABLE_TEST_ID } from './test_ids';
 import type { CtiEnrichment } from '../../../../../common/search_strategy';
-import { QUERY_ID } from '../../shared/hooks/use_investigation_enrichment';
+import { QUERY_ID } from '../../../../flyout_v2/document/hooks/use_investigation_enrichment';
 import { InspectButton } from '../../../../common/components/inspect';
 import {
   buildThreatDetailsItems,
