+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/insights_section.tsx
@@ -9,11 +9,13 @@ import { type DataTableRecord, getFieldValue } from '@kbn/discover-utils';
 import { i18n } from '@kbn/i18n';
 import React, { memo, useCallback, useMemo } from 'react';
 import { EVENT_KIND } from '@kbn/rule-data-utils';
+import { EventKind } from '../constants/event_kinds';
 import { useRuleWithFallback } from '../../../detection_engine/rule_management/logic/use_rule_with_fallback';
 import { FLYOUT_STORAGE_KEYS } from '../constants/local_storage';
 import { PREFIX } from '../../../flyout/shared/test_ids';
 import { ExpandableSection } from '../../shared/components/expandable_section';
 import { useExpandSection } from '../../shared/hooks/use_expand_section';
+import { ThreatIntelligenceOverview } from './threat_intelligence_overview';
 import { PrevalenceOverview } from './prevalence_overview';
 
 export const INSIGHTS_SECTION_TEST_ID = `${PREFIX}InsightsSection` as const;
@@ -45,7 +47,10 @@ export const InsightsSection = memo(({ hit }: InsightsSectionProps) => {
     defaultValue: true,
   });
 
-  const isAlert = useMemo(() => (getFieldValue(hit, EVENT_KIND) as string) === 'signal', [hit]);
+  const isAlert = useMemo(
+    () => (getFieldValue(hit, EVENT_KIND) as string) === EventKind.signal,
+    [hit]
+  );
   const ruleId = useMemo(
     () =>
       isAlert
@@ -59,6 +64,7 @@ export const InsightsSection = memo(({ hit }: InsightsSectionProps) => {
     [rule?.investigation_fields?.field_names]
   );
 
+  const onShowThreatIntelligenceDetails = useCallback(() => {}, []);
   const onShowPrevalenceDetails = useCallback(() => {}, []);
 
   return (
@@ -70,6 +76,13 @@ export const InsightsSection = memo(({ hit }: InsightsSectionProps) => {
       sectionId={LOCAL_STORAGE_SECTION_KEY}
       title={INSIGHTS_SECTION_TITLE}
     >
+      {isAlert && (
+        <ThreatIntelligenceOverview
+          hit={hit}
+          onShowThreatIntelligence={onShowThreatIntelligenceDetails}
+          showIcon={false}
+        />
+      )}
       <PrevalenceOverview
         hit={hit}
         investigationFields={investigationFields}
