+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/test_ids.ts
@@ -48,4 +48,10 @@ export const SESSION_VIEW_NO_DATA_TEST_ID = `${PREFIX}SessionViewNoData` as cons
 
 /* Insights */
 
+export const INSIGHTS_THREAT_INTELLIGENCE_TEST_ID = `${PREFIX}InsightsThreatIntelligence` as const;
+export const INSIGHTS_THREAT_INTELLIGENCE_THREAT_MATCHES_TEST_ID =
+  `${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}ThreatMatches` as const;
+export const INSIGHTS_THREAT_INTELLIGENCE_ENRICHED_WITH_THREAT_INTELLIGENCE_TEST_ID =
+  `${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}EnrichedWithThreatIntelligence` as const;
+
 export const PREVALENCE_TEST_ID = `${PREFIX}InsightsPrevalence` as const;
