+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/hooks/use_fetch_threat_intelligence.ts
@@ -1,0 +1,96 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { useMemo } from 'react';
+import { groupBy } from 'lodash';
+import type { DataTableRecord } from '@kbn/discover-utils';
+import type { CtiEnrichment } from '../../../../common/search_strategy';
+import {
+  filterDuplicateEnrichments,
+  getEnrichmentFields,
+  parseExistingEnrichments,
+} from '../utils/threat_intelligence_helpers';
+import { useInvestigationTimeEnrichment } from './use_investigation_enrichment';
+import { ENRICHMENT_TYPES } from '../../../../common/cti/constants';
+
+export interface UseThreatIntelligenceParams {
+  /**
+   * Document record used to retrieve threat intelligence data
+   */
+  hit: DataTableRecord;
+}
+
+export interface UseThreatIntelligenceResult {
+  /**
+   * Returns true while the threat intelligence data is being queried
+   */
+  loading: boolean;
+  /**
+   * Threat matches (from an indicator match rule)
+   */
+  threatMatches: CtiEnrichment[];
+  /**
+   * Threat matches count
+   */
+  threatMatchesCount: number;
+  /**
+   * Threat enrichments (from the real time query)
+   */
+  threatEnrichments: CtiEnrichment[];
+  /**
+   *  Threat enrichments count
+   */
+  threatEnrichmentsCount: number;
+}
+
+/**
+ * Hook to retrieve threat intelligence data for the expandable flyout right and left sections.
+ */
+export const useFetchThreatIntelligence = ({
+  hit,
+}: UseThreatIntelligenceParams): UseThreatIntelligenceResult => {
+  const alertRuleUuid = hit.flattened['kibana.alert.rule.uuid'];
+  const isAlert = Array.isArray(alertRuleUuid)
+    ? alertRuleUuid.some((value) => value != null)
+    : alertRuleUuid != null;
+
+  // retrieve the threat enrichment fields with value for the current document
+  // (see https://github.com/elastic/kibana/blob/main/x-pack/solutions/security/plugins/security_solution/common/cti/constants.ts#L35)
+  const eventFields = useMemo(() => getEnrichmentFields(hit), [hit]);
+
+  // retrieve existing enrichment fields and their value
+  const existingEnrichments = useMemo(
+    () => (isAlert ? parseExistingEnrichments(hit) : []),
+    [hit, isAlert]
+  );
+
+  // api call to retrieve all documents that match the eventFields
+  const { result: response, loading } = useInvestigationTimeEnrichment({ eventFields });
+
+  // combine existing enrichment and enrichment from the api response
+  // also removes the investigation-time enrichments if the exact indicator already exists
+  const allEnrichments = useMemo(() => {
+    if (loading || !response?.enrichments) {
+      return existingEnrichments;
+    }
+    return filterDuplicateEnrichments([...existingEnrichments, ...response.enrichments]);
+  }, [loading, response, existingEnrichments]);
+
+  // separate threat matches (from indicator-match rule) from threat enrichments (realtime query)
+  const {
+    [ENRICHMENT_TYPES.IndicatorMatchRule]: threatMatches,
+    [ENRICHMENT_TYPES.InvestigationTime]: threatEnrichments,
+  } = groupBy(allEnrichments, 'matched.type');
+
+  return {
+    loading,
+    threatMatches,
+    threatMatchesCount: (threatMatches || []).length,
+    threatEnrichments,
+    threatEnrichmentsCount: (threatEnrichments || []).length,
+  };
+};
